Rename funky packages and interface
