+++ src/java/magmac/app/compile/error/type/package-info.java
@@ -5,4 +5,4 @@
  * Recommendation: Small leaf package; no immediate need to break up or merge.
  * </p>
  */
-package magmac.app.compile.error.error;
+package magmac.app.compile.error.type;
