+++ src/java/magmac/app/compile/rule/ContextRule.java
@@ -3,7 +3,7 @@ package magmac.app.compile.rule;
 import magmac.api.collect.list.Lists;
 import magmac.app.compile.error.CompileResult;
 import magmac.app.compile.error.context.NodeContext;
-import magmac.app.compile.error.error.CompileError;
+import magmac.app.compile.error.type.CompileError;
 import magmac.app.compile.node.Node;
 import magmac.app.compile.error.context.StringContext;
 import magmac.app.error.ImmutableCompileError;
