+++ src/java/magmac/app/compile/rule/FilterRule.java
@@ -1,7 +1,7 @@
 package magmac.app.compile.rule;
 
 import magmac.app.compile.error.CompileResult;
-import magmac.app.compile.error.error.CompileErrors;
+import magmac.app.compile.error.type.CompileErrors;
 import magmac.app.compile.node.Node;
 import magmac.app.compile.rule.filter.Filter;
 import magmac.app.compile.rule.filter.NumberFilter;
