+++ src/java/magmac/app/compile/rule/TypeRule.java
@@ -5,8 +5,8 @@ import magmac.app.compile.error.CompileResult;
 import magmac.app.compile.error.context.Context;
 import magmac.app.compile.error.context.NodeContext;
 import magmac.app.compile.error.context.StringContext;
-import magmac.app.compile.error.error.CompileError;
-import magmac.app.compile.error.error.CompileErrors;
+import magmac.app.compile.error.type.CompileError;
+import magmac.app.compile.error.type.CompileErrors;
 import magmac.app.compile.node.Node;
 import magmac.app.error.ImmutableCompileError;
